chatbot/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
openai==1.78.0
python-dotenv==1.1.0
scikit-learn==1.6.1
numpy==2.2.5
